tests/requirements.txt
# Test Dependencies for Monk-AI Comprehensive Test Cycle

# Core testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# HTTP client libraries
aiohttp>=3.8.0
requests>=2.28.0
httpx>=0.24.0

# AI Provider SDKs
openai>=1.0.0
google-generativeai>=0.3.0

# FastAPI and related
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Data handling
json5>=0.9.0
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.0.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Performance monitoring
psutil>=5.9.0
memory-profiler>=0.60.0

# Reporting
jinja2>=3.1.0
markdown>=3.4.0